[ADD] 메서드별 상세 권한추가(전략, 관리자)

+++ src/main/java/com/sysmatic2/finalbe/strategy/service/StrategyReviewService.java
@@ -19,6 +19,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -112,6 +113,11 @@ public class StrategyReviewService {
         MemberEntity updater = memberRepository.findById(updaterId)
                 .orElseThrow(() -> new MemberNotFoundException("Updater not found with ID: " + updaterId));
 
+        //리뷰 수정은 작성자만 가능하다.
+        if(!review.getUpdaterId().equals(updaterId)){
+            throw new AccessDeniedException("리뷰 수정은 작성자만 가능합니다.");
+        }
+
         // 수정
         review.setUpdaterId(updater);
         review.setContent(updatedContent);
@@ -131,11 +137,33 @@ public class StrategyReviewService {
      * 5-1. 리뷰 삭제
      */
     @Transactional
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(Long strategyId, Long reviewId, String memberId, Boolean isAdmin, Boolean isTrader) {
         StrategyReviewEntity review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new ReviewNotFoundException("Review not found with ID: " + reviewId));
 
-        reviewRepository.delete(review);
+        //관리자인 경우
+        if(isAdmin){
+            reviewRepository.delete(review);
+            return;
+        }
+
+        //전략 정보
+        StrategyEntity strategyEntity = strategyRepository.findById(strategyId)
+                        .orElseThrow(() -> new StrategyNotFoundException("Strategy not found with ID: " + strategyId + "in review "));
+
+        //트레이더인 경우 : 전략의 작성자
+        if(isTrader && strategyEntity.getWriterId().equals(memberId)){
+            reviewRepository.delete(review);
+            return;
+        }
+
+        //일반 유저인 경우 : 리뷰의 작성자
+        if(review.getWriterId().equals(memberId)){
+            reviewRepository.delete(review);
+            return;
+        }
+
+        throw new AccessDeniedException("리뷰 삭제 권한이 없습니다.");
     }
 
     /**
